// packages/fe-job-pc/src/pages/admin/service/hierarchical-diagnosis/diagnosis/interface.d.ts
export interface SurveyScoreItem {
    value?: number
}

export interface JudgmentResultItem {
    value?: number
}

export interface DimensionHierarchyItem {
    value?: number
}